Two bug fixes: Go back to preferring Device::SerialPort, and only setup serial connection if needed
It turns out that plain I/O as I'm using it now has some problems
that are avoided by using Device::SerialPort, so I'm going back
to using that for now.

It was not possible to use bc125at-perl driver on a system that
lacked /dev/ttyUSB0, because the serial connection setup died
when that was not found. This has been fixed by only setting up
the serial connection if needed.