Add optional setuid helper program to automate driver setup.
This "driver helper" is basically just a C reimplementation of
Bc125At::Detect, which may be installed setuid in /usr/local/bin
in order to give bc125at-perl (particularly the GUI) a way to
easily redetect the scanner and reload the usbserial driver. This
is a fairly common thing to have to do, for example, during initial
setup, after unplugging the device and plugging it back in, or after
rebooting.