Initial implementation of unattended "auto scan" feature.
The goal of this feature (which is still in its very early stages)
is to allow the software to control the scanner while listening for
audio, and keep a record of which channels had audio present at which
times of day. This data could be used to rank channels from most
interesting to least interesting (to avoid wasting time on boring
channels), to generate fancy graphs, or for other purposes.