Implement auto search feature.
The purpose of this feature is to facilitate systematic (but slow)
discovery of new frequencies to monitor. The main advantage is that
you can leave it running for extended periods of time and then
view the report when it's done. A disadvantage is the extremely
slownes compared to the built-in search feature. At the moment,
it can only hit about one frequency per second, so don't hold your
breath waiting to find a hit. The search order can also be
randomized in order to make the process slightly less boring to
watch.