Workaround for "cdc_acm 3-4:1.0: Zero length descriptor references".
Issue 1: https://github.com/rikus--/bc125at-perl/issues/1

Due to an apparent limitation of the Uniden hardware, this device, at
least on more recent kernels, needs the NO_UNION_NORMAL quirk.
A kernel patch was submitted to automatically apply this quirk for the
BC125XLT, but the kernel is not patched for the BC125AT yet.

Based on a suggestion found on a mailing list, I found that the kernel
patch is not necessary after all if you want to apply a simple workaround
to the already-running system.

See: https://www.spinics.net/lists/linux-usb/msg158942.html